+++ packages/desktop/src/index.ts
@@ -18,6 +18,8 @@ export * from './list'
 export * from './pagination'
 export * from './header-menu'
 export * from './header'
+export * from './status'
+export * from './logo-link'
 
 export {TextField as MaskTextField} from './text-field'
 export {PasswordField as MaskPasswordField} from './password-field'