+++ packages/desktop/src/pagination/Pagination.tsx
@@ -1,7 +1,7 @@
 import React, {FunctionComponent} from 'react'
 
-import {Box, Card, IconWrapper, PaginationControl} from '@qiwi/pijma-core'
-import {Icon} from '@qiwi/pijma-media'
+import {PaginationControl, Card, Icon} from '@qiwi/pijma-core'
+
 import {Text} from '../typography'
 import {PaginationLink} from './PaginationLink'
 
@@ -30,8 +30,6 @@ export const Pagination: FunctionComponent<PaginationProps> = ({
     onChange={onChange}
     children={renderProps => (
       <Card
-        role="toolbar"
-        aria-label="pagination"
         bg="#fff"
         s={shadowed ? '0 1px 2px 0 rgba(0,0,0,0.12)' : '0 0 0 1px #e6e6e6'}
         height={12}
@@ -47,11 +45,7 @@ export const Pagination: FunctionComponent<PaginationProps> = ({
           href={href ? href(1) : undefined}
           onClick={renderProps.onPageClick(1, renderProps.pages.includes(1))}
           children={({disabled}) => (
-            <Box width={5} height={5}>
-              <IconWrapper color={disabled ? '#ccc' : '#000'}>
-                <Icon name="angle-double-left-small" />
-              </IconWrapper>
-            </Box>
+            <Icon name="angle-double-left-small" color={disabled ? '#ccc' : '#000'}/>
           )}
         />
         <PaginationLink
@@ -61,11 +55,7 @@ export const Pagination: FunctionComponent<PaginationProps> = ({
           href={href ? href(renderProps.prev) : undefined}
           onClick={renderProps.onPageClick(renderProps.prev, renderProps.pages.includes(1))}
           children={({disabled}) => (
-            <Box width={5} height={5}>
-              <IconWrapper color={disabled ? '#ccc' : '#000'}>
-                <Icon name="angle-small-left" />
-              </IconWrapper>
-            </Box>
+            <Icon name="angle-small-left" color={disabled ? '#ccc' : '#000'}/>
           )}
         />
         {renderProps.pages.map(page => (
@@ -81,9 +71,8 @@ export const Pagination: FunctionComponent<PaginationProps> = ({
                 bold={false}
                 size="m"
                 color={renderProps.active === page ? 'warning' : 'default'}
-              >
-                {page}
-              </Text>
+                children={page}
+              />
             )}
           />
         ))}
@@ -94,11 +83,7 @@ export const Pagination: FunctionComponent<PaginationProps> = ({
           href={href ? href(renderProps.next) : undefined}
           onClick={renderProps.onPageClick(renderProps.next, renderProps.pages.includes(renderProps.total))}
           children={({disabled}) => (
-            <Box width={5} height={5}>
-              <IconWrapper color={disabled ? '#ccc' : '#000'}>
-                <Icon name="angle-small-right" />
-              </IconWrapper>
-            </Box>
+            <Icon name="angle-small-right" color={disabled ? '#ccc' : '#000'}/>
           )}
         />
         <PaginationLink
@@ -109,11 +94,7 @@ export const Pagination: FunctionComponent<PaginationProps> = ({
           href={href ? href(renderProps.total) : undefined}
           onClick={renderProps.onPageClick(renderProps.total, renderProps.pages.includes(renderProps.total))}
           children={({disabled}) => (
-            <Box width={5} height={5}>
-              <IconWrapper color={disabled ? '#ccc' : '#000'}>
-                <Icon name="angle-double-right-small" />
-              </IconWrapper>
-            </Box>
+            <Icon name="angle-double-right-small" color={disabled ? '#ccc' : '#000'}/>
           )}
         />
       </Card>