+++ packages/mobile/src/checkbox-field/CheckboxField.tsx
@@ -25,7 +25,6 @@ const CheckboxField: FunctionComponent<
     onBlur={props.onBlur}
     children={(renderProps) => (
       <OptionField
-        role="checkbox"
         title={props.title}
         hint={props.hint}
         help={props.help}
@@ -44,7 +43,6 @@ const CheckboxField: FunctionComponent<
             onMouseLeave={renderProps.onMouseLeave}
             children={(renderOptionProps) => (
               <OptionFieldItem
-                role="menuitemcheckbox"
                 disabled={option.disabled}
                 icon={(
                   <Box width={6}>