+++ packages/mobile/src/typography/Title.tsx
@@ -8,6 +8,7 @@ export interface TitleProps {
   color?: 'default' | 'inverse'
   align?: TypoProps['align']
   stub?: boolean
+  clamp?: number
 }
 
 const TitleSize: { [size in TitleProps['size']]: number } = {
@@ -50,7 +51,7 @@ const TitleColor: { [color in NonNullable<TitleProps['color']>]: string } = {
   inverse: '#fff',
 }
 
-export const Title: FunctionComponent<TitleProps> = ({tag, size, color = 'default', align, stub, children}) => (
+export const Title: FunctionComponent<TitleProps> = ({tag, size, color = 'default', align, stub, clamp, children}) => (
   stub ? (
     <Box
       ml={align === 'center' || align === 'right' ? 'auto' : 'none'}
@@ -74,6 +75,7 @@ export const Title: FunctionComponent<TitleProps> = ({tag, size, color = 'defaul
       weight={TitleWeight[size]}
       color={TitleColor[color]}
       align={align}
+      clamp={clamp}
       children={<Breaker children={children}/>}
     />
   )